requirements.txt
matplotlib==3.8.2
natsort==8.4.0
numpy==1.26.4
opencv-python==4.8.1.78
pandas==2.1.3
pillow==9.5.0
scikit-image==0.24.0
thop==0.1.1.post2209072238
torchvision==0.16.1
tqdm==4.66.1
warmup-scheduler==0.3
